Rename sac utils
