+++ torchbenchmark/models/pytorch_unet/install.py
@@ -1,4 +1,4 @@
 from utils.python_utils import pip_install_requirements
 
 if __name__ == '__main__':
-    pip_install_requirements()
+    pip_install_requirements(requirements_txt="pytorch_unet/requirements.txt")
